### 2025-12-25 785406a
Version 1.1.1 : adding citations and authors
