+++ paper.bib
@@ -4,7 +4,8 @@
   journal={Transportation Research Part B-methodological},
   year={2002},
   volume={36},
-  pages={507-535}
+  pages={507-535},
+  doi = {https://doi.org/10.1016/S0191-2615(01)00015-7},
 }
 
 
@@ -16,6 +17,7 @@
   SERIES = {Modeling and Simulation in Science, Engineering and Technology},
   YEAR = {2023},
   KEYWORDS = {pedestrian  flow ; Hughes model ; conservation law ; moving interface ; many-particle approximation ; wave-front tracking},
+  doi = {https://doi.org/10.1007/978-3-031-46359-4_2}
 }
 
 @article{LW55,
@@ -26,7 +28,8 @@
   number = {1178},
   pages = {317--345},
   year = {1955},
-  organization = {The Royal Society}
+  organization = {The Royal Society},
+  doi = {https://doi.org/10.1098/rspa.1955.0089}
 }
 
 @article{Ric56,
@@ -37,7 +40,8 @@
   number = {1},
   pages = {42--51},
   year = {1956},
-  publisher = {INFORMS}
+  publisher = {INFORMS},
+  doi = {https://doi.org/10.1287/opre.4.1.42}
 }
 
 @article{CGLM11,
@@ -47,6 +51,7 @@
   year={2011},
   volume={22},
   pages={1150023},
+  doi = {https://doi.org/10.1142/s0218202511500230}
 }
 
 @phdthesis{Gir25,
@@ -56,6 +61,8 @@ conservation scalaires rencontrées dans les modèles de dynamique piétonnière
 school = {Université de Tours},
 type = {PhD thesis},
 year = {2025},
+URL = {https://hal.science/tel-05275359},
+HAL_ID = {tel-05275359},
 }
 
 @article{AbrahamPreprint,
@@ -69,7 +76,8 @@ VOLUME = {30},
 PUBLISHER = {Springer, Cham},
 title = {Finite volume approximation and well-posedness of conservation laws with moving interfaces under abstract coupling conditions},
 KEYWORDS = {Conservation laws, Discontinuous flux, Interface coupling conditions, Numerical
-scheme}
+scheme},
+doi = {https://doi.org/10.1007/s00030-023-00857-9}
 }
 
 @article{Goatin2014,
@@ -81,6 +89,7 @@ year = {2014},
 note = {The Conference on Pedestrian and Evacuation Dynamics 2014 (PED 2014), 22-24 October 2014, Delft, The Netherlands},
 issn = {2352-1465},
 author = {Monika Twarogowska and Paola Goatin and Regis Duvigneau},
+doi = {https://doi.org/10.1016/j.trpro.2014.09.063}
 }
 
 @software{vadere,
@@ -88,19 +97,22 @@ author = {Benedikt Kleinmeier, Benedikt Zönnchen, Marion Gödel, Gerta Köster}
 title = {Vadere: An Open-Source Simulation Framework to Promote Interdisciplinary Understanding.},
 journal = {Collective Dynamics, vol. 4},
 year = {2019},
+doi = {https://doi.org/10.17815/cd.2019.21}
 }
 
 @software{JuPedSim,
 author = {Chraibi, Mohcine and Kratz, Kai and Schrödter, Tobias and {The JuPedSim Development Team}},
 title = {{JuPedSim}},
-url = {https://github.com/PedestrianDynamics/jupedsim},
-version = {v1.1.0}
+repository-code = {https://github.com/PedestrianDynamics/jupedsim},
+version = {v1.1.0},
+url = {https://jupedsim.org}
 }
 
 @software{UMANS,
[…]
