+++ paper.md
@@ -12,7 +12,7 @@ authors:
     orcid: 0000-0002-9382-6746
     affiliation: 1
 affiliations:
-  - name: Institut Denis Poisson, Université de Tours
+  - name: Institut Denis Poisson, Université de Tours, France
     index: 1
 date: 02 december 2025
 bibliography: paper.bib
@@ -57,25 +57,10 @@ A classical example of such a speed function is $v(\rho) = v_{max}\frac{\rho_{ma
 ## The Eikonal equation
 
 The second equation of the model characterizes the unitary direction field $\vec{V}(t,x)$ depending on the density $\rho$ in the whole domain. We assume that the pedestrians want to minimize their exit time while also trying to avoid high density regions. In order to model this situation, we use an optimal control problem. We suppose that the density $\rho(\cdot) \in \mathcal{C}^1(\bar \Omega)$ stays constant in time (this assumption is quite controversial).
-Let $x \in \Omega$. For any $\alpha(\cdot) \in \mathcal{C}^1((0,+\infty),\mathcal{S}_1)$, we say that $X^\alpha_x(\cdot)$ is a trajectory controlled by $\alpha$ starting at $x$ if $X$ is a solution to the Cauchy problem:
+Then, for any $x \in \Omega$, we define the value function by
 
-
-$$
- \left\lbrace \begin{matrix}
- \dot{X}^\alpha_x(s) = \alpha(s)v(\rho(t,X^\alpha_x(s))) \\
- X^\alpha_x(0) = x
- \end{matrix}\right.
- $$
-
-For any $x \in \Omega$, we denote by $\mathcal{A}_x = \lbrace X^\alpha_x, \alpha \in \mathcal{C}^1((0,+\infty),\mathcal{S}_1) \rbrace$ the set of all controlled trajectories starting at $x$. We define $\phi(x)$ as the minimal exit time starting at location $x$, that is to say:
-
-
-$$ \phi(x) = \inf_{X \in \mathcal{A}_x} \int_0^{+\infty} \mathbb{1}_{\Omega}(X(t)) \textrm{d} t.$$
-
-In Hughes' model, we also take into account the discomfort caused by being surrounded by a high density crowd. In order to model this discomfort, we introduce an increasing function $g(\rho)$ with respect to the density $\rho$. The function $g(\rho)$ can be interpreted as a running cost we are paying along a trajectory $X$ for being in high density regions. Then, the previous equation becomes:
-
-
-$$ \phi(x) = \inf_{X \in \mathcal{A}_x} \int_0^{+\infty} \mathbb{1}_{\Omega}(X(t))g(\rho(X(t))) \textrm{d} t.$$
+$$ \phi(x) = \inf_{X \in \mathcal{A}_x} \int_0^{+\infty} \mathbb{1}_{\Omega}(X(t))g(\rho(X(t))) \textrm{d} t,$$
+where $A_x$ denotes the set of admissible trajectories and $g$ is an increasing function modeling the discomfort of a agent standing in a high density region.
 
 A very classical result of the theory of viscosity solution for Hamilton-Jacobi-Bellman (HJB) equations is that
 solving the optimal control problem above is in fact equivalent to solving the Eikonal equation:
@@ -115,34 +100,11 @@ $$
 
 For a deep overview of the mathematical results regarding Hughes' model we defer to [@survey].
 
+> **Remark:** We also included a numerical scheme for different models in the Hughes2d package such as the model of Colombo-Garavello-Lécureux-Mercier (see [@CGLM11]) but for conciseness' sake, we only present the Hughes model.
 
-## Model of Colombo-Garavello-Lécureux-Mercier
+# Software design
 
-In [@CGLM11], the authors introduced an alternative model for pedestrian flows.
-
[…]
