+++ paper.bib
@@ -5,7 +5,7 @@
   year={2002},
   volume={36},
   pages={507-535},
-  doi = {https://doi.org/10.1016/S0191-2615(01)00015-7},
+  doi = {10.1016/S0191-2615(01)00015-7},
 }
 
 
@@ -17,7 +17,7 @@
   SERIES = {Modeling and Simulation in Science, Engineering and Technology},
   YEAR = {2023},
   KEYWORDS = {pedestrian  flow ; Hughes model ; conservation law ; moving interface ; many-particle approximation ; wave-front tracking},
-  doi = {https://doi.org/10.1007/978-3-031-46359-4_2}
+  doi = {10.1007/978-3-031-46359-4_2}
 }
 
 @article{LW55,
@@ -29,7 +29,7 @@
   pages = {317--345},
   year = {1955},
   organization = {The Royal Society},
-  doi = {https://doi.org/10.1098/rspa.1955.0089}
+  doi = {10.1098/rspa.1955.0089}
 }
 
 @article{Ric56,
@@ -41,7 +41,7 @@
   pages = {42--51},
   year = {1956},
   publisher = {INFORMS},
-  doi = {https://doi.org/10.1287/opre.4.1.42}
+  doi = {10.1287/opre.4.1.42}
 }
 
 @article{CGLM11,
@@ -51,7 +51,7 @@
   year={2011},
   volume={22},
   pages={1150023},
-  doi = {https://doi.org/10.1142/s0218202511500230}
+  doi = {10.1142/s0218202511500230}
 }
 
 @phdthesis{Gir25,
@@ -77,7 +77,7 @@ PUBLISHER = {Springer, Cham},
 title = {Finite volume approximation and well-posedness of conservation laws with moving interfaces under abstract coupling conditions},
 KEYWORDS = {Conservation laws, Discontinuous flux, Interface coupling conditions, Numerical
 scheme},
-doi = {https://doi.org/10.1007/s00030-023-00857-9}
+doi = {10.1007/s00030-023-00857-9}
 }
 
 @article{Goatin2014,
@@ -89,7 +89,7 @@ year = {2014},
 note = {The Conference on Pedestrian and Evacuation Dynamics 2014 (PED 2014), 22-24 October 2014, Delft, The Netherlands},
 issn = {2352-1465},
 author = {Monika Twarogowska and Paola Goatin and Regis Duvigneau},
-doi = {https://doi.org/10.1016/j.trpro.2014.09.063}
+doi = {10.1016/j.trpro.2014.09.063}
 }
 
 @software{vadere,
@@ -97,7 +97,7 @@ author = {Benedikt Kleinmeier, Benedikt Zönnchen, Marion Gödel, Gerta Köster}
 title = {Vadere: An Open-Source Simulation Framework to Promote Interdisciplinary Understanding.},
 journal = {Collective Dynamics, vol. 4},
 year = {2019},
-doi = {https://doi.org/10.17815/cd.2019.21}
+doi = {10.17815/cd.2019.21}
 }
 
 @software{JuPedSim,
@@ -112,7 +112,7 @@ url = {https://jupedsim.org}
 title = {Generalized Microscopic Crowd Simulation using Costs in Velocity Space},
 author = {Wouter van Toll, Fabien Grzeskowiak, Axel López, Javad Amirian, Florian Berton, Julien Bruneau, Beatriz Cabrero Daniel, Alberto Jovane, Julien Pettré},
 year = {2020},
-doi = {https://doi.org/10.1145/3384382.3384532}
+doi = {10.1145/3384382.3384532}
 }
 
 @book{Cromosim,
[…]
