### 2026-01-26 #9608 @TheoRGirard
@danielskatz I corrected most of the issues pointed by the editorial bot !

I only reduced the word count to 1338 words... Would that be okay taking into account the additional sections ?

Once again sorry for the late answer !
