### 2026-01-26 #9608 @danielskatz
Thanks @TheoRGirard - note that "The paper should be between 750-1750 words." - https://joss.readthedocs.io/en/latest/paper.html  We've increased the word could because of the extra sections we now require.

I'll check this more carefully in the next few days.
