### 2026-01-28 #9608 @danielskatz
Hi - We've found a small inconsistency in our documentation and requirements associated with our recent scope chanage, which we've corrected, and I apologize for asking you to do more work, but could you be sure you have a State of the Field (~200 words) section in your paper?

A description of how this software compares to other commonly-used packages in the research area. If related tools exist, provide a clear "build vs. contribute" justification explaining your unique scholarly contribution and why existing alternatives are insufficient.

For full details, please see:
* Our updated scope definition: https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance
* An example paper in the new format: https://joss.readthedocs.io/en/latest/example_paper.html
