### 2026-01-29 #9608 @TheoRGirard
@danielskatz Hi ! I've checked and modified a bit the structure of the paper !
I'm now pretty convinced it satisfies the required structure but don't hesitate to tell me if you detect any issue !
