### 2026-02-06 #9608 @slayoo
> Hi [@slayoo](https://github.com/slayoo) do you have time to review this paper?

I'm in!
