### 2026-02-23 #10106 @editorialbot
[REVIEW]: hughes2d: approximating solutions of Hughes’ model
**Submitting author:** @TheoRGirard (<a href="http://orcid.org/0000-0002-9382-6746">Théo Girard</a>)
**Repository:** https://github.com/TheoRGirard/hughes2d
**Branch with paper.md** (empty if default branch): main
**Version:** 1.1.0
