### 2026-03-30 eb5fc4b
paper text suggestions, typo fixes, etc
see https://github.com/openjournals/joss-reviews/issues/10106
