+++ paper.md
@@ -1,7 +1,7 @@
 ---
-title: "Hughes2d: approximating solutions of Hughes' model"
+title: "hughes2d: Approximating solutions to the Hughes pedestrian-flow model"
 tags:
-  - python
+  - Python
   - non-linear discontinuous PDE
   - pedestrian evacuation
   - macroscopic crowd dynamics
@@ -20,47 +20,47 @@ bibliography: paper.bib
 
 # Summary
 
-`hughes2d` is an open-source python package for simulation pedestrian crowds in two dimensions. More specifically, the package is designed to compute approximations of Hughes' model introduced in [@Hug02] (and other related models). The Hughes model is a macroscopic model -there is no agents here, the crowd is represented by a density function- coupling two non-linear partial differential equations.
+`hughes2d` is an open-source python package for simulation pedestrian crowds in two dimensions. More specifically, the package is designed to compute approximations of the Hughes model introduced in [@Hug02] (and other related models). The Hughes model is a macroscopic model -there is no agents here, the crowd is represented by a density function- coupling two non-linear partial differential equations.
 
 
 # State of the field
 
-The mathematical modeling of pedestrian crowd is a rapidly developping topic since a few decades. There exist multiple software for simulation crowds of pedestrians both open source ([@vadere],[@JuPedSim],[@UMANS],[@Cromosim],[@CrowdWalk]) or not (PTV viswalk, MassMotion). We defer to the [awesome-crowdynamics repository](https://github.com/pozapas/awesome-crowdynamics) for an exhaustive list of the available softwares. However, up to our knowledge, all these softwares deal with microscopic simulations. We propose here a python package for macroscopic simulations of pedestrian evacuations, specifically for Hughes' model which is one of the most famous macroscopic pedestrian flow models.
+The mathematical modeling of pedestrian crowd is a rapidly developing topic since a few decades. There exist multiple software packages for simulating crowds of pedestrians -- both open source [@vadere;@JuPedSim;@UMANS;@Cromosim;@CrowdWalk]) and proprietary (PTV viswalk, MassMotion). We defer to the [awesome-crowdynamics repository](https://github.com/pozapas/awesome-crowdynamics) for an exhaustive list of the available packages. However, to the best of our knowledge, all these developments deal with microscopic simulations. We propose here a Python package for macroscopic simulations of pedestrian evacuations, specifically for the seminal Hughes pedestrian flow model.
 
 
 # Statement of need
 
-The Hughes model has been thoroughly studied during the last two decades (see [@survey]) but there exists, at the moment, no general mathematical result of existence of solutions in 2D for this model. Some simulations appear in a few papers (see [@Goatin2014]) but in a slightly modified context. The present package should provide a reliable and open-source solution to approximate the behavior of Hughes' model.
+The Hughes model has been thoroughly studied during the last two decades [see @survey], but there exists, at the moment, no general mathematical result of existence of solutions in 2D for this model. While example numerical solutions to the Hughes model can be found in literature (e.g., [@Goatin2014]), these are not supported by a publicly available and reusable codebase. The present package is aimed at providing a reliable and open-source solution to approximate the behavior of the Hughes model.
 We also hope that this package will help formulating conjectures in the future.
 
 
 # Software design
 
[…]
