### 2026-03-30 15569e8
Fix formatting issues in paper.bib entries
