+++ paper.bib
@@ -10,7 +10,7 @@
 
 
 @incollection{survey,
-  TITLE = {{The mathematical theory of Hughes' model : a survey of results}},
+  TITLE = {{The mathematical theory of Hughes' model: a survey of results}},
   AUTHOR = {Amadori, Debora and Andreianov, Boris and Di Francesco, Marco and Fagioli, Simone and Girard, Th{\'e}o and Goatin, Paola and Markowich, Peter and Pietschmann, Jan F. and Rosini, Massimiliano D. and Russo, Giovanni and Stivaletta, Graziano and Wolfram, Marie-Therese},
   BOOKTITLE = {{Crowd Dynamics, Volume 4 : Analytics and Human Factors in Crowd Modeling}},
   PUBLISHER = {{Springer}},
@@ -56,7 +56,7 @@
 
 @phdthesis{Gir25,
 author = {Théo Girard},
-title = {Discontinuités dans les équations d’Hamilton-Jacobi-Bellman et dans les lois de
+title = {Discontinuités dans les équations d’{H}amilton-{J}acobi-{B}ellman et dans les lois de
 conservation scalaires rencontrées dans les modèles de dynamique piétonnière},
 school = {Université de Tours},
 type = {PhD thesis},
@@ -93,8 +93,8 @@ doi = {10.1016/j.trpro.2014.09.063}
 }
 
 @software{vadere,
-author = {Benedikt Kleinmeier, Benedikt Zönnchen, Marion Gödel, Gerta Köster},
-title = {Vadere: An Open-Source Simulation Framework to Promote Interdisciplinary Understanding.},
+author = {Benedikt Kleinmeier and Benedikt Zönnchen and Marion Gödel and Gerta Köster},
+title = {Vadere: An Open-Source Simulation Framework to Promote Interdisciplinary Understanding},
 journal = {Collective Dynamics, vol. 4},
 year = {2019},
 doi = {10.17815/cd.2019.21}
