### 2026-03-30 #10106 @slayoo
- [ ] in https://github.com/TheoRGirard/hughes2d/pull/1, I've proposed a few suggestions to the text:
  - title: JOSS paper archives reveal that projects with all-lowercase names, retain this capitalisation in the titles,
    thus suggesting to change from "Hughes2d: approximating..." to "hughes2d: Approximating" (although, I've noticed that in the README, while `hughes2d` is used in the text, the animation actually uses capital "H" and capital "D")
  - suggest to consistently use "Hughes model" instead of "Hughes' model" - in the same way we write about "Stokes flow" rather than "Stokes’ flow" (originally, both options where used throughout the text)
  - "software" is uncountable in English
  - "python" -> "Python"
  - there are multiple instances in the paper where nested or adjacent parentheses were used, like "((" or ")(" - best to avoid these
  - the introduction features contradicting statements indicating that there are no publicly available codes for the problem, yet at the same time stating "most of the prototypes of code produced by mathematicians work only on square meshes" - the later statement would require a reference to be retained.
  - avoiding italics in subscripts constituting labels rather than values (e.g., \rho_{max} -> \rho_{\max})
  - consistent "eikonal" capitalisation (originally both Eikonal and eikonal were used - please pick one)
  - upright font for physical units
  - capitalisation in references (e.g., hamilton-jacobi-bellman)
  - typos

Furthermore:
- [ ] suggest to persistently archive the videos mentioned from the text at Zenodo, in addition to GitHub
- [ ] figure quality is not sufficient for publication, please increase label sizes to make them readable, and use vector graphics instead of PNG format
- [ ] despite highlighting in the introduction that "there is no agents here" (fixed in the above PR to "there are"), the paper repeatedly uses the "agent" jargon, and even points that the figures depict that "agents seem to be separated in two different groups" - suggest to elaborate

These are my comments for the text, I'm now moving to checking out the code aiming at reproducing the figure myself.
