### 2026-03-30 #10106 @slayoo
Here are some first notes:
- [ ] the README does not explain how to run the tests (since Python has numerous testing frameworks, just mentioning that `pytest` is used here might be relevant)
- [ ] suggest to switch from manual version setting in `pyproject.toml` to something like `setuptools_scm` which would ensure it is consistent with git tags and corresponding releases;
- [ ] `import hughes2d; print(hughes2d.__version__)` should not err
- [ ] is there a reason to maintain (autogenerated) `requirements.txt` file in the repo, if the data inside is also present in `pyproject.toml`?
- [ ] suggest to set up Zenodo.org release archival (see https://help.zenodo.org/docs/github/)
- [ ] suggest to set up CI workflow with uploads to PyPI.org (also to ensure that the project name will not be used by another project in future)
- [ ] running the tests on Python 3.13 results in one warning:
```
pytest
======================================== test session starts ==============================================
rootdir: /home/user/devel/hughes2d
configfile: pyproject.toml
collected 61 items

test/test_Mesh2D_importExport.py ...                                                                                                                                                                                                  [  4%]
test/test_Mesh2D_squareDomain.py ...........................                                                                                                                                                                          [ 49%]
test/test_Mesh2D_squareDomain_plots.py ...                                                                                                                                                                                            [ 54%]
test/test_run_examples.py .......                                                                                                                                                                                                     [ 65%]
test/test_solvers_LWR.py .....                                                                                                                                                                                                        [ 73%]
test/test_solvers_eikonal.py ...........                                                                                                                                                                                              [ 91%]
test/test_solvers_pedestrian.py .....                                                                                                                                                                                                 [100%]

======================================= warnings summary ==============================================
test/test_solvers_pedestrian.py::test_pedestrian_models[hughes-0.1-100-0.04]
test/test_solvers_pedestrian.py::test_pedestrian_models[colombo-garavello-0.1-100-0.09]
test/test_solvers_pedestrian.py::test_pedestrian_models[constantDirectionField-0.1-100-0.09]
  /home/user/devel/hughes2d/src/hughes2d/Splitting.py:206: DeprecationWarning: datetime.datetime.utcnow() is deprecated and scheduled for removal in a future version. Use timezone-aware objects to represent datetimes in UTC: datetime.datetime.now(datetime.UTC).
    +str(datetime.utcnow().date()))

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
================================== 61 passed, 3 warnings in 48.21s =========================================
```
