+++ paper.bib
@@ -81,7 +81,7 @@ doi = {10.1007/s00030-023-00857-9}
 }
 
 @article{Goatin2014,
-title = {Comparative Study of Macroscopic Pedestrian Models},
+title = {Comparative study of macroscopic pedestrian models},
 journal = {Transportation Research Procedia},
 volume = {2},
 pages = {477-485},
@@ -93,8 +93,13 @@ doi = {10.1016/j.trpro.2014.09.063}
 }
 
 @software{vadere,
+<<<<<<< HEAD
 author = {Benedikt Kleinmeier and Benedikt Zönnchen and Marion Gödel and Gerta Köster},
 title = {Vadere: An Open-Source Simulation Framework to Promote Interdisciplinary Understanding},
+=======
+author = {Benedikt Kleinmeier, Benedikt Zönnchen, Marion Gödel, Gerta Köster},
+title = {Vadere: an open-source simulation framework to promote interdisciplinary understanding.},
+>>>>>>> e1874a3 (Fixing paper from slayoo's review)
 journal = {Collective Dynamics, vol. 4},
 year = {2019},
 doi = {10.17815/cd.2019.21}
@@ -109,7 +114,7 @@ url = {https://jupedsim.org}
 }
 
 @software{UMANS,
-title = {Generalized Microscopic Crowd Simulation using Costs in Velocity Space},
+title = {Generalized microscopic crowd simulation using costs in velocity space},
 author = {Wouter van Toll, Fabien Grzeskowiak, Axel López, Javad Amirian, Florian Berton, Julien Bruneau, Beatriz Cabrero Daniel, Alberto Jovane, Julien Pettré},
 year = {2020},
 doi = {10.1145/3384382.3384532}
