### 2026-04-14 4ebf284
Temporarily fixing python-package.yml because coverage-badge is broken
