+++ .github/workflows/python-package.yml
@@ -27,7 +27,9 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install flake8 pytest pytest-cov coverage-badge setuptools
+        python -m pip install flake8 pytest pytest-cov coverage-badge 
+        #temporary fix since coverage-badge depends on pkg_resource that is no longer included in setuptools >= 82.0.0
+        python -m pip install 'setuptools<82.0.0' --force-reinstall 
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
         pip install -e .
     - name: Lint with flake8
