+++ src/hughes2d/_version.py
@@ -1,0 +1,24 @@
+# file generated by vcs-versioning
+# don't change, don't track in version control
+from __future__ import annotations
+
+__all__ = [
+    "__version__",
+    "__version_tuple__",
+    "version",
+    "version_tuple",
+    "__commit_id__",
+    "commit_id",
+]
+
+version: str
+__version__: str
+__version_tuple__: tuple[int | str, ...]
+version_tuple: tuple[int | str, ...]
+commit_id: str | None
+__commit_id__: str | None
+
+__version__ = version = '0.1.dev224+gc3990b0c4.d20260414'
+__version_tuple__ = version_tuple = (0, 1, 'dev224', 'gc3990b0c4.d20260414')
+
+__commit_id__ = commit_id = 'gc3990b0c4'
