+++ src/hughes2d.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+hughes2d
