+++ src/hughes2d.egg-info/SOURCES.txt
@@ -1,0 +1,90 @@
+.coverage
+.gitignore
+.readthedocs.yaml
+AUTHORS
+CHANGELOG.md
+CONTRIBUTING.md
+LICENCE
+README.md
+paper.bib
+paper.md
+pyproject.toml
+requirements.txt
+uv.lock
+.github/workflows/draft-pdf.yml
+.github/workflows/python-coverage.yml
+.github/workflows/python-package.yml
+docs/Makefile
+docs/make.bat
+docs/requirements.txt
+docs/source/api.rst
+docs/source/conf.py
+docs/source/gettingStartedVid.gif
+docs/source/index.rst
+docs/source/maths.rst
+docs/source/reference.rst
+docs/source/usage.rst
+docs/source/assets/CompareExplicit.png
+docs/source/assets/MidVectorSchema.png
+docs/source/assets/constant_circle.png
+docs/source/assets/demo.png
+docs/source/assets/mesh_generated.png
+docs/source/assets/square_domain.png
+examples/00-getting_started/getting_started.py
+examples/01-Mesh_generation/Generating_square_mesh.py
+examples/02-Simulating_Hughes_1/Simulating_Grandmont_Restaurant.py
+examples/02-Simulating_Hughes_1/data/Mp4converter.py
+examples/03-Simulating_Hughes_2/Comparison_Hughes_vs_explicit_vector_field.py
+examples/04-Simulating_Eikonal_equation/Comparing_numerical_methods.py
+examples/04-Simulating_Eikonal_equation/Simulating_shortest_path.py
+examples/04-Simulating_Eikonal_equation/data/square_mesh.json
+examples/05-Simulating_LWR/Simulating_LWR.py
+examples/06-File_handling/FreeFem/TestExportMSH.py
+examples/06-File_handling/FreeFem/TestImportMSH.py
+examples/06-File_handling/FreeFem/mesh_FF.msh
+examples/06-File_handling/dxf/DXFImport.py
+examples/06-File_handling/dxf/config_simple.dxf
+examples/06-File_handling/dxf/config_simple_zones.dxf
+examples/06-File_handling/dxf/definingZones.py
+examples/06-File_handling/msh/ImportMSH.py
+examples/06-File_handling/msh/insulated-2.2.msh
+examples/06-File_handling/vtk/mesh_FF.vtk
+examples/07-Additional_Computations/AdditionalComputations.py
+examples/08-Plot_methods/time_slices/TimeSlices.py
+examples/08-Plot_methods/time_slices/time_slices_densities.csv
+examples/08-Plot_methods/time_slices/time_slices_mesh.json
+examples/08-Plot_methods/video/Hughes_densities.csv
+examples/08-Plot_methods/video/Hughes_mesh.json
+examples/08-Plot_methods/video/convert.py
+logo/logo.gif
+logo/logo.mp4
+src/hughes2d/EikonalSolver.py
+src/hughes2d/LWR2D.py
+src/hughes2d/Mesh2D.py
+src/hughes2d/Plotter.py
+src/hughes2d/Splitting.py
+src/hughes2d/__init__.py
+src/hughes2d/_version.py
+src/hughes2d.egg-info/PKG-INFO
+src/hughes2d.egg-info/SOURCES.txt
+src/hughes2d.egg-info/dependency_links.txt
+src/hughes2d.egg-info/entry_points.txt
+src/hughes2d.egg-info/requires.txt
+src/hughes2d.egg-info/top_level.txt
+test/__init__.py
+test/test_Mesh2D_importExport.py
+test/test_Mesh2D_squareDomain.py
+test/test_Mesh2D_squareDomain_plots.py
+test/test_run_examples.py
[…]
