+++ src/hughes2d.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+hughes2d = hughes2d:main
