### 2026-04-14 #10106 @TheoRGirard
> [@TheoRGirard](https://github.com/TheoRGirard) could you please have a look at the notes above?

@diehlpk I'm currently working on the notes of @slayoo ! Most of the remarks on the paper have been readily applied.
The warning is fixed and I will upload to Zenodo as soon as I look into it, I didn't know the existence of the database.

Regarding the automatic version management, integration issues, upload on PyPI and the requirements.txt remark, I'm trying to find a reliable solution. I'm guessing that the main issue is that I'm not properly using the versioning tools from uv as I'm not sufficiently familiar with it. Regarding the requirements.txt, the file is needed because I'm using pip in github actions instead of uv. Once again, this should change and it would be ideal to setup the continuous integration with uv instead of pip.
