+++ paper.bib
@@ -93,13 +93,8 @@ doi = {10.1016/j.trpro.2014.09.063}
 }
 
 @software{vadere,
-<<<<<<< HEAD
 author = {Benedikt Kleinmeier and Benedikt Zönnchen and Marion Gödel and Gerta Köster},
 title = {Vadere: An Open-Source Simulation Framework to Promote Interdisciplinary Understanding},
-=======
-author = {Benedikt Kleinmeier, Benedikt Zönnchen, Marion Gödel, Gerta Köster},
-title = {Vadere: an open-source simulation framework to promote interdisciplinary understanding.},
->>>>>>> e1874a3 (Fixing paper from slayoo's review)
 journal = {Collective Dynamics, vol. 4},
 year = {2019},
 doi = {10.17815/cd.2019.21}
