### 2026-04-14 d9ca649
Deleting wrong integration of setuptools_scm
