+++ src/hughes2d.egg-info/requires.txt
@@ -1,15 +1,0 @@
-numpy>=2.2.2
-triangle>=20250106
-
-[file-formats]
-ezdxf>=1.3.5
-meshio>=5.3.5
-
-[file-import]
-
-[msh-files]
-
-[plot]
-matplotlib>=3.10.0
-plotly>=6.0.0
-pyqt6>=6.8.1
