+++ pyproject.toml
@@ -1,6 +1,11 @@
 [project]
 name = "hughes2d"
+<<<<<<< HEAD
 version = "1.1.3"
+=======
+version = "1.1.2"
+#dynamic = ["version"]
+>>>>>>> 90db695 (Deleting wrong integration of setuptools_scm)
 description = "A python package for simulating macroscopic pedestrian evacuations."
 readme = "README.md"
 authors = [
@@ -32,13 +37,9 @@ file-import = []
 requires = ["setuptools>=80", "setuptools-scm[simple]>=9.2"]
 build-backend = "setuptools.build_meta"
 
-[tool.setuptools_scm]
-version_file = "src/hughes2d/_version.py"
-
 [dependency-groups]
 dev = [
     "coverage>=7.8.2",
     "pytest>=8.3.4",
     "pytest-cov>=6.1.1",
-    "setuptools-scm>=10.0.5",
 ]
