+++ .github/workflows/run_tests.yml
@@ -27,7 +27,7 @@ jobs:
         python-version: ${{ matrix.python-version }}
 
     - name: Install uv
-        uses: astral-sh/setup-uv@v7
+      uses: astral-sh/setup-uv@v7
 
     - name: Install dependencies without extras
       run: |
