+++ .github/workflows/run_tests.yml
@@ -17,7 +17,10 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.11", "3.12", "3.13"]
+        python-version:
+          - "3.11"
+          - "3.12"
+          - "3.13"
 
     steps:
     - uses: actions/checkout@v4
@@ -28,6 +31,8 @@ jobs:
 
     - name: Install uv
       uses: astral-sh/setup-uv@v7
+      with:
+          python-version: ${{ matrix.python-version }}
 
     - name: Install dependencies without extras
       run: |
