+++ .github/workflows/run_tests.yml
@@ -2,7 +2,7 @@
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 # And for example of an integration with uv see: https://github.com/pyjanitor-devs/pyjanitor/tree/dev/.github/workflows
 
-name: Python package
+name: Tests with pytest
 
 on:
   push:
