### 2026-04-15 bb99b02
Delete .github/workflows/python-coverage.yml
