+++ CHANGELOG.md
@@ -1,4 +1,3 @@
-# Change log
 
 ## [1.1.2] - (14/04/2026)
 
