+++ .github/workflows/run_tests.yml
@@ -37,18 +37,15 @@ jobs:
     - name: Install dependencies without extras
       run: |
         uv sync
-        uv add pytest coveralls
+        uv add pytest pytest-cov coverage-badge
         
-    - name: Test with pytest and generate coverage bade
+    - name: Test with pytest and generate .coverage
       run: |
-        uv run -m pytest
+        uv run coverage run -m pytest
 
-    - name: Run Coverage
+    - name: Run coverage-badge
       run: |
-        uv run -m coverage run -m pytest
+        #temporary fix since coverage-badge depends on pkg_resource that is no longer included in setuptools >= 82.0.0
+        uv run --with 'setuptools<82.0.0' coverage-badge -o coverage.svg
         
-    - name: Run coveralls
-      run: coveralls
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         
