### 2026-04-16 d3edaec
Update Comparison_Hughes_vs_explicit_vector_field.py
