+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "hughes2d"
-version = "1.1.2"
+version = "1.1.3"
 description = "A python package for simulating macroscopic pedestrian evacuations."
 readme = "README.md"
 authors = [
