### 2026-04-16 38d112b
Setup of pre-commit
