+++ pyproject.toml
@@ -31,6 +31,7 @@ file-import = []
 [dependency-groups]
 dev = [
     "coverage>=7.8.2",
+    "pre-commit>=4.5.1",
     "pytest>=8.3.4",
     "pytest-cov>=6.1.1",
 ]
