+++ pyproject.toml
@@ -28,6 +28,10 @@ file-formats = [
 msh-files = []
 file-import = []
 
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
 [dependency-groups]
 dev = [
     "coverage>=7.8.2",
