+++ requirements.txt
@@ -1,9 +1,0 @@
-# This file was autogenerated by uv via the following command:
-#    uv pip compile pyproject.toml -o requirements.txt
-numpy>=2.2.2
-    # via
-    #   hughes2d (pyproject.toml)
-    #   triangle
-triangle>=20250106
-    # via hughes2d (pyproject.toml)
-typing_extensions>=4.0.1
