+++ pyproject.toml
@@ -3,39 +3,30 @@ name = "hughes2d"
 version = "1.1.3"
 description = "A python package for simulating macroscopic pedestrian evacuations."
 readme = "README.md"
-authors = [
-    { name = "Theo Girard", email = "theo.girard@math.cnrs.fr" }
-]
 requires-python = ">=3.11"
-dependencies = [
-    "numpy>=2.2.2",
-    "triangle>=20250106",
-]
+dependencies = [ "numpy>=2.2.2", "triangle>=20250106" ]
 
-[project.scripts]
-hughes2d = "hughes2d:main"
+  [[project.authors]]
+  name = "Theo Girard"
+  email = "theo.girard@math.cnrs.fr"
 
-[project.optional-dependencies]
-plot = [
-    "matplotlib>=3.10.0",
-    "plotly>=6.0.0",
-    "pyqt6>=6.8.1",
-]
-file-formats = [
-    "ezdxf>=1.3.5",
-    "meshio>=5.3.5",
-]
-msh-files = []
-file-import = []
+  [project.scripts]
+  hughes2d = "hughes2d:main"
+
+  [project.optional-dependencies]
+  plot = [ "matplotlib>=3.10.0", "plotly>=6.0.0", "pyqt6>=6.8.1" ]
+  file-formats = [ "ezdxf>=1.3.5", "meshio>=5.3.5" ]
+  msh-files = [ ]
+  file-import = [ ]
 
 [build-system]
-requires = ["hatchling"]
+requires = [ "hatchling" ]
 build-backend = "hatchling.build"
 
 [dependency-groups]
 dev = [
-    "coverage>=7.8.2",
-    "pre-commit>=4.5.1",
-    "pytest>=8.3.4",
-    "pytest-cov>=6.1.1",
+  "coverage>=7.8.2",
+  "pre-commit>=4.5.1",
+  "pytest>=8.3.4",
+  "pytest-cov>=6.1.1"
 ]
