+++ pyproject.toml
@@ -1,3 +1,4 @@
+
 [project]
 name = "hughes2d"
 version = "1.1.3"
@@ -24,9 +25,4 @@ requires = [ "hatchling" ]
 build-backend = "hatchling.build"
 
 [dependency-groups]
-dev = [
-  "coverage>=7.8.2",
-  "pre-commit>=4.5.1",
-  "pytest>=8.3.4",
-  "pytest-cov>=6.1.1"
-]
+dev = [ "coverage>=7.8.2", "pre-commit>=4.5.1", "pytest>=8.3.4", "pytest-cov>=6.1.1" ]
