### 2026-04-16 959fbdb
Create auto-release.yml
