+++ .github/workflows/auto-release.yml
@@ -1,0 +1,77 @@
+name: Auto-release
+
+on:
+  workflow_dispatch:
+
+permissions:
+  contents: write
+  id-token: write
+
+jobs:
+  release:
+    name: Create a new release
+    runs-on: ubuntu-latest
+
+    defaults:
+      run:
+        shell: bash -l {0}
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v6
+        with:
+          fetch-depth: 0
+          fetch-tags: true
+
+      - name: Check if already released
+        id: check_release
+        run: |
+          CURRENT_COMMIT=$(git rev-parse HEAD)
+          echo "Current commit: $CURRENT_COMMIT"
+          if git tag --points-at $CURRENT_COMMIT | grep -q "^v"; then
+            echo "already_released=true" >> $GITHUB_OUTPUT
+            echo "This commit already has a release tag. Will skip release."
+            git tag --points-at $CURRENT_COMMIT
+          else
+            echo "already_released=false" >> $GITHUB_OUTPUT
+            echo "No release tag found. Proceeding with release."
+          fi
+
+      - name: Setup Python environment
+        if: steps.check_release.outputs.already_released == 'false'
+        uses: actions/setup-python@v6
+        with:
+          python-version: 3.13
+
+      - name: Install uv
+        if: steps.check_release.outputs.already_released == 'false'
+        uses: astral-sh/setup-uv@v7
+
+      - name: Ensure repo status is clean
+        if: steps.check_release.outputs.already_released == 'false'
+        run: git status
+
+      - name: Build package
+        if: steps.check_release.outputs.already_released == 'false'
+        run: uv build
+
+      - name: Publish package (trusted publishing)
+        if: steps.check_release.outputs.already_released == 'false'
+        run: uv publish --trusted-publishing always
+
+      - name: Get version number from uv
+        if: steps.check_release.outputs.already_released == 'false'
+        run: echo "VERSION_NUM = $(uv version --short)" >> $GITHUB_OUTPUT
+        id: version
+
+      - name: Create release in GitHub repo
+        if: steps.check_release.outputs.already_released == 'false'
+        uses: ncipollo/release-action@v1
+        with:
+          bodyFile: "docs/releases/v${{ steps.version.outputs.VERSION_NUM }}.md"
+          token: ${{ secrets.GITHUB_TOKEN }}
+          tag: v${{ steps.version.outputs.VERSION_NUM }}
+
+      - name: Ensure complete
+        if: steps.check_release.outputs.already_released == 'false'
+        run: echo "Auto-release complete!"
