+++ .github/workflows/auto-release.yml
@@ -12,10 +12,6 @@ jobs:
     name: Create a new release
     runs-on: ubuntu-latest
 
-    defaults:
-      run:
-        shell: bash -l {0}
-
     steps:
       - name: Checkout repository
         uses: actions/checkout@v6
