### 2026-04-16 fab08c8
Update auto-release.yml
