+++ .github/workflows/auto-release.yml
@@ -51,10 +51,6 @@ jobs:
         if: steps.check_release.outputs.already_released == 'false'
         run: uv build
 
-      - name: Publish package (trusted publishing)
-        if: steps.check_release.outputs.already_released == 'false'
-        run: uv publish --trusted-publishing always
-
       - name: Get version number from uv
         if: steps.check_release.outputs.already_released == 'false'
         run: echo "VERSION_NUM = $(uv version --short)" >> $GITHUB_OUTPUT
@@ -67,6 +63,10 @@ jobs:
           bodyFile: "docs/releases/v${{ steps.version.outputs.VERSION_NUM }}.md"
           token: ${{ secrets.GITHUB_TOKEN }}
           tag: v${{ steps.version.outputs.VERSION_NUM }}
+
+      - name: Publish package (trusted publishing)
+        if: steps.check_release.outputs.already_released == 'false'
+        run: uv publish --trusted-publishing always
 
       - name: Ensure complete
         if: steps.check_release.outputs.already_released == 'false'
