### 2026-04-16 4318fac
Update auto-release.yml
