+++ .github/workflows/auto-release.yml
@@ -53,8 +53,13 @@ jobs:
 
       - name: Get version number from uv
         if: steps.check_release.outputs.already_released == 'false'
-        run: echo "VERSION_NUM = $(uv version --short)" >> $GITHUB_OUTPUT
+        run: echo "VERSION_NUM = $(uv version --short | tr -d ' \n')" >> $GITHUB_OUTPUT
         id: version
+
+      - name: Display new version number
+        if: steps.check_release.outputs.already_released == 'false'
+        run: |
+          echo "version_number: v${{ steps.version.outputs.VERSION_NUM }}"
 
       - name: Create release in GitHub repo
         if: steps.check_release.outputs.already_released == 'false'
