### 2026-04-16 6c06ce0
Update auto-release.yml
