+++ .github/workflows/auto-release.yml
@@ -53,21 +53,20 @@ jobs:
 
       - name: Get version number from uv
         if: steps.check_release.outputs.already_released == 'false'
-        run: echo "VERSION_NUM = $(uv version --short | tr -d ' \n')" >> $GITHUB_OUTPUT
-        id: version
+        run: echo "version_number=$(uv version --short | tr -d ' \n')" >> $GITHUB_ENV
 
       - name: Display new version number
         if: steps.check_release.outputs.already_released == 'false'
         run: |
-          echo "version_number: v${{ steps.version.outputs.VERSION_NUM }}"
+          echo "version_number: v${{ env.version_number }}"
 
       - name: Create release in GitHub repo
         if: steps.check_release.outputs.already_released == 'false'
         uses: ncipollo/release-action@v1
         with:
-          bodyFile: "docs/releases/v${{ steps.version.outputs.VERSION_NUM }}.md"
+          bodyFile: "docs/releases/v${{ env.version_number }}.md"
           token: ${{ secrets.GITHUB_TOKEN }}
-          tag: v${{ steps.version.outputs.VERSION_NUM }}
+          tag: v${{ env.version_number }}
 
       - name: Publish package (trusted publishing)
         if: steps.check_release.outputs.already_released == 'false'
