### 2026-04-16 5f2a4ca
Update auto-release.yml
