+++ paper.bib
@@ -137,3 +137,15 @@ month     =  mar,
 year      =  2013,
 doi = {10.9746/jcmsi.6.137}
 }
+
+@misc{videoGirardHughes26,
+  author       = {Girard, Théo},
+  title        = {A few macroscopic pedestrian crowd simulations
+                   with the hughes2d package.
+                  },
+  month        = apr,
+  year         = 2026,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.19613348},
+  url          = {https://doi.org/10.5281/zenodo.19613348},
+}
