### 2026-04-16 7cee165
Changing figures to svg
