### 2026-04-16 1245e28
Setting back the figure as a png
