+++ src/hughes2d/__init__.py
@@ -15,4 +15,4 @@ __all__ = [
         "VertexValueMap",
         ]
 
-__version__ = '1.1.3'
+__version__ = '1.1.4'
