+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-- Setup of auto-releasing workflow with publishing to pyPI
+
+## [1.1.4] - (16/04/2026)
+
+- Incorporating the JOSS reviews and comments - Setup of auto-releasing workflow with publishing to pyPI
 - Fixing et reworking paper
 - Setup of pre-commit
 - Deleting requirements.txt
