### 2026-04-16 #10106 @TheoRGirard
I've corrected all of the issues raised by @slayoo apart from the inclusion of .svg files in the draft. I have uploaded the corresponding svg files in the repo (at docs/source/assets/demo/) but I did not find a way to directly include svg in the PDF file. In markdown, I can use HTML but the <img> doesn't seem to import the image in the pdf.

Otherwise, I have now a fully automatic versioning, releasing and publishing workflow (but I chose to trigger it manually with a workflow dispatch because of the small scale of the project).

Thanks a lot for pointing all these issues and for correcting the draft paper !
