### 2026-05-11 a33239d
Update logo image link in README.md (hope this will fix display on https://pypi.org/p/hughes2d)
