### 2026-05-11 4ff07c4
Revise README installation instructions for pip (the `requirements.txt` file is no longer in the repo; the `-e` option requires an argument)
