### 2026-05-15 6e6f2a3
explain how to run tess in README.md
Updated testing instructions in README.md to include pytest documentation and usage.
