### 2026-05-15 37fadb8
cleanup CONTRIBUTING.md removing one broken link, and a TOC for a file with just a few sentences
Removed unnecessary sections and cleaned up formatting.
