### 2026-05-19 f880b73
Fixing the docstring and tests for VertexValuedMap plot methods
