+++ src/hughes2d/Mesh2D.py
@@ -21,21 +21,27 @@ from pathlib import Path
 import numpy as np
 import triangle as tr
 
+# for Type checking
+from collections.abc import Callable
+from typing import Self
+
+from numpy.typing import ArrayLike
+
 FigureType = None
 AxesType = None
 
 try:
     import matplotlib.pyplot as plt
     from matplotlib import cm, collections, patches
-    #FigureType = FigureType | plt.Figure
-    #AxesType = AxesType | plt.Axes
+    # FigureType = FigureType | plt.Figure
+    # AxesType = AxesType | plt.Axes
 except ImportError:
     plt = None
 
 try:
     import plotly.figure_factory as ff
     import plotly.graph_objects as go
-    #FigureType = FigureType | go.Figure
+    # FigureType = FigureType | go.Figure
 except ImportError:
     go, ff = None, None
 
@@ -54,19 +60,13 @@ try:
 except ImportError:
     json = None
 
-#for Type checking
-from collections.abc import Callable
-from typing import Self
-
-from numpy.typing import ArrayLike
-
 PointType = list[float]
 
 PRECISION = 1e-10
-DEFAULT_MESH_DICT = {"domain" : 0, "exit" : [98], "wall" : [99]}
-PLACEHOLDER_DOMAIN = [[0,0],[1,0],[0,1]]
+DEFAULT_MESH_DICT = {"domain": 0, "exit": [98], "wall": [99]}
+PLACEHOLDER_DOMAIN = [[0, 0], [1, 0], [0, 1]]
 DEFAULT_REQUIREMENTS = ["all"]
-DEFAULT_COLOR_RGBA = [244,22,100,0.6]
+DEFAULT_COLOR_RGBA = [244, 22, 100, 0.6]
 TWO = 2
 
 
@@ -119,11 +119,13 @@ class NonConvexDomain:
     --------
     """
 
-    def __init__(self, outer_vertices_list:list[PointType] = PLACEHOLDER_DOMAIN) -> None:
+    def __init__(
+        self, outer_vertices_list: list[PointType] = PLACEHOLDER_DOMAIN
+    ) -> None:
         self.outer_vertices = outer_vertices_list
         self.outer_boundary = []
         for i in range(len(outer_vertices_list)):
-            self.outer_boundary.append([i,(i+1)%(len(outer_vertices_list))])
+            self.outer_boundary.append([i, (i + 1) % (len(outer_vertices_list))])
         self.wall_vertices = []
         self.wall_edges = []
         self.wall_holes_point = []
@@ -165,16 +167,16 @@ class NonConvexDomain:
 
         # entity query for all LINE entities in modelspace
         for e in msp.query("LINE"):
-            if(e.dxf.layer not in {"0","domain"}):
+            if e.dxf.layer not in {"0", "domain"}:
                 continue
-            point1 = [np.round(e.dxf.start[0],5),np.round(e.dxf.start[1],5)]
+            point1 = [np.round(e.dxf.start[0], 5), np.round(e.dxf.start[1], 5)]
             if point1 not in self.outer_vertices:
                 id_point1 = len(self.outer_vertices)
                 self.outer_vertices.append(point1)
             else:
                 id_point1 = self.add_boundary_point(point1)
 
-            point2 = [np.round(e.dxf.end[0],5),np.round(e.dxf.end[1],5)]
+            point2 = [np.round(e.dxf.end[0], 5), np.round(e.dxf.end[1], 5)]
             if point2 not in self.outer_vertices:
                 id_point2 = len(self.outer_vertices)
                 self.outer_vertices.append(point2)
@@ -183,68 +185,66 @@ class NonConvexDomain:
 
             self.outer_boundary.append([id_point1, id_point2])
 
[…]
