### 2026-05-19 525bcf3
Increasing fontsize in save_time_slices method
