+++ src/hughes2d/Plotter.py
@@ -26,22 +26,29 @@ FigureType = None
 try:
     import matplotlib.pyplot as plt
     from matplotlib import animation, cm, collections
-    #FigureType = FigureType | plt.Figure
+    # FigureType = FigureType | plt.Figure
 except ImportError:
     plt = None
 
 try:
     import plotly.figure_factory as ff
     import plotly.graph_objects as go
-    #FigureType = FigureType | go.Figure
+    # FigureType = FigureType | go.Figure
 except ImportError:
     go, ff = None, None
 
 
-
 EMPTY_LIST = []
 
-def convert_to_mp4(filename:str, limits:list[list[float]] = EMPTY_LIST, dpi_set:int = 300, pic_size:tuple[int,int] = (8,6), *, plot_scale:bool = True) -> None:
+
+def convert_to_mp4(
+    filename: str,
+    limits: list[list[float]] = EMPTY_LIST,
+    dpi_set: int = 300,
+    pic_size: tuple[int, int] = (8, 6),
+    *,
+    plot_scale: bool = True,
+) -> None:
     """Create a video file from the data files of a simulation.
 
     Args:
@@ -61,22 +68,23 @@ def convert_to_mp4(filename:str, limits:list[list[float]] = EMPTY_LIST, dpi_set:
         msg = "matplotlib is required for this function."
         raise ImportError(msg)
 
-    with Path(filename +"_mesh.json").open("r") as f:
+    with Path(filename + "_mesh.json").open("r") as f:
         data = json.load(f)
-        triangles = np.array([ [ [ data["vertices"][i][0], data["vertices"][i][1] ]
-                                for i in triangle] for triangle in data["triangles"] ])
-
+        triangles = np.array(
+            [
+                [[data["vertices"][i][0], data["vertices"][i][1]] for i in triangle]
+                for triangle in data["triangles"]
+            ]
+        )
 
     values = []
 
-    with Path(filename +"_densities.csv").open("r") as file:
+    with Path(filename + "_densities.csv").open("r") as file:
         csv_file = csv.reader(file)
         for lines in csv_file:
-            values = [*values,
-                      np.array([float(lines[i]) for i in range(len(lines))])]
+            values = [*values, np.array([float(lines[i]) for i in range(len(lines))])]
 
-
-    fig, ax = plt.subplots(figsize= pic_size, dpi=dpi_set)
+    fig, ax = plt.subplots(figsize=pic_size, dpi=dpi_set)
 
     col = collections.PolyCollection(triangles)
     col.set_cmap(cm.viridis)
@@ -87,27 +95,34 @@ def convert_to_mp4(filename:str, limits:list[list[float]] = EMPTY_LIST, dpi_set:
     if plot_scale:
         fig.colorbar(rgcol, ax=ax, label="density")
 
-
     if len(limits) > 0:
-        ax.set_xlim(limits[0][0],limits[0][1])
-        ax.set_ylim(limits[1][0],limits[1][1])
+        ax.set_xlim(limits[0][0], limits[0][1])
+        ax.set_ylim(limits[1][0], limits[1][1])
         plt.axis("equal")
 
-
-    def update(frame:int):
+    def update(frame: int):
         col.set_array(values[frame])
         ax.add_collection(col)
 
-        return (col,ax)
-
+        return (col, ax)
 
     ani = animation.FuncAnimation(fig=fig, func=update, frames=len(values), interval=30)
 
     ff_writer = animation.FFMpegWriter(fps=25)
-    ani.save(filename+".mp4", writer = ff_writer, dpi=dpi_set)
+    ani.save(filename + ".mp4", writer=ff_writer, dpi=dpi_set)
 
 
[…]
