### 2026-05-20 4e28496
Update auto-release.yml
