### 2026-05-20 5d8c7f9
Updatinh the paper's figure and large-file precommit size to 900
