+++ .pre-commit-config.yaml
@@ -21,6 +21,7 @@ repos:
     -   id: check-yaml
     -   id: check-toml
     -   id: check-added-large-files
+        args: ['--maxkb=900']
 
 -   repo: https://github.com/astral-sh/uv-pre-commit
     # uv version.
