### 2026-05-20 264865c
Updatinh the paper's figure
