### 2026-05-20 41e305a
Updating the paper's figure
