### 2026-06-30 1fc9ae1
Update paper.md
