### 2026-06-30 5eb301b
Multiprocessing windows-compatible implementation
