+++ src/hughes2d/Splitting.py
@@ -29,11 +29,15 @@ from hughes2d.EikonalSolver import EikoSolver
 from hughes2d.LWR2D import LWRSolver
 from hughes2d.Mesh2D import CellValueMap, Mesh, VertexValueMap
 
+if __name__ == "__main__":
+    multiprocessing.freeze_support()
+
 previous_process_dens = object()
 previous_process_vec = object()
 EMPTY_LIST = []
-DEFAULT_OPTIONS = {"model":"hughes"}
+DEFAULT_OPTIONS = {"model": "hughes"}
 EMPTY_THRESHOLD = 1e-2
+
 
 class PedestrianSolver:
     """An object wrapping together different models for pedestrian flows.
@@ -169,67 +173,73 @@ class PedestrianSolver:
     -------
     """
 
-    def __init__(self,
-                 mesh:Mesh,
-                 dt:float,
-                 initial_density:CellValueMap,
-                 speed_function:Callable[[float],float] = (lambda x: 1-x),
-                 cost_function:Callable[[float],float] = (lambda x: 1+2*x),
-                 directions:list[list[float]] = EMPTY_LIST,
-                 options:dict=DEFAULT_OPTIONS) -> None:
+    def __init__(
+        self,
+        mesh: Mesh,
+        dt: float,
+        initial_density: CellValueMap,
+        speed_function: Callable[[float], float] = (lambda x: 1 - x),
+        cost_function: Callable[[float], float] = (lambda x: 1 + 2 * x),
+        directions: list[list[float]] = EMPTY_LIST,
+        options: dict = DEFAULT_OPTIONS,
+    ) -> None:
         self.mesh = mesh
         self.time_step = 0
 
         self.options = options
 
-        if("framerate" not in self.options):
+        if "framerate" not in self.options:
             self.options["framerate"] = 25
 
         self.dt = dt
 
-        if("verbose" not in self.options):
+        if "verbose" not in self.options:
             self.options["verbose"] = False
 
-        self.num_forgotten_steps = max(int(1/(self.dt*self.options["framerate"])), 1)
+        self.num_forgotten_steps = max(
+            int(1 / (self.dt * self.options["framerate"])), 1
+        )
         if self.options["verbose"]:
             print("Number of steps omitted for one frame : ", self.num_forgotten_steps)
-
 
         self.speed_function = speed_function
         self.cost_function = cost_function
 
-        if("save" not in self.options):
+        if "save" not in self.options:
             self.options["save"] = False
 
-        if("filename" not in self.options):
-            self.options["filename"] = ("Save"
-                                        +str(datetime.datetime.now(datetime.UTC)))
+        if "filename" not in self.options:
+            self.options["filename"] = "Save" + str(datetime.datetime.now(datetime.UTC))
 
-        if("additional_computations" not in self.options):
+        if "additional_computations" not in self.options:
             self.options["additional_computations"] = {}
 
-        if("model" not in self.options):
+        if "model" not in self.options:
             msg = "A model should be prescribed in the options dictionary."
             raise ValueError(msg)
 
-        if("eikoSolver" not in self.options):
-            self.options["eikoSolver"] = {  "constrained" : True,
-                                            "NarrowBandDepth" : 2}
+        if "eikoSolver" not in self.options:
+            self.options["eikoSolver"] = {"constrained": True, "NarrowBandDepth": 2}
 
-        if("lwrSolver" not in self.options):
-            self.options["lwrSolver"] = {   "convexFlux" : True,
-                                            "anNum" : "dichotomy",
-                                            "method" : "midVector",
-                                            "ApproximationThreshold" : 0.0001}
+        if "lwrSolver" not in self.options:
+            self.options["lwrSolver"] = {
+                "convexFlux": True,
[…]
