+++ pyproject.toml
@@ -30,4 +30,9 @@ requires = [ "hatchling" ]
 build-backend = "hatchling.build"
 
 [dependency-groups]
-dev = [ "coverage>=7.8.2", "pre-commit>=4.5.1", "pytest>=8.3.4", "pytest-cov>=6.1.1" ]
+dev = [
+ "coverage>=7.8.2",
+ "pre-commit>=4.5.1",
+ "pytest>=8.3.4",
+ "pytest-cov>=6.1.1",
+]
