### 2026-07-01 1f885f6
Removing useless old code
