### 2026-07-01 417cc4f
Revert Plotter.py
