### 2026-07-01 e293eb2
Making the fontsize adjustable in save_time_slices
