### 2026-07-01 2ba173f
Updating demo figure
