### 2026-07-01 ecd7bd2
Update paper.md
