### 2026-07-01 aac81e7
Allow temporarily use of Node20 in setup_ffmpeg in run_tests.yml
