### 2026-07-01 57c6078
Update paper.md
