+++ paper.md
@@ -123,7 +123,7 @@ On a mathematical point of view, the simulation using the Hughes model suggests 
 Apart from the mathematical conjectures we derive from the simulations, we also hope that, by making an open source software, more researchers will take an interest in macroscopic pedestrian models. A numerical comparison of the simulation using `hughes2d` with other models is already on-going (see [BOUM project](https://conferences.cirm-math.fr/3512.html)).
 Many of the tools developed here for approximating Hughes model are not specific to the PDEs involved. Most of the codebase could be easily reused and incorporated in other macroscopic pedestrian crowd projects.
 
-![Sample simulation using the Hughes model](docs/source/assets/demo/figure_demo.pdf)
+![Sample simulation of a pedestrian density using the Hughes model in the 2D domain $\bar\Omega$](docs/source/assets/demo/figure_demo.pdf)
 
 # AI usage disclosure
 
