+++ examples/08-Plot_methods/video/convert.py
@@ -4,4 +4,4 @@ import hughes2d.Plotter
 
 filename = str(Path(__file__).parent / "Hughes")
 
-hughes2d.Plotter.convertToMP4(filename)
+hughes2d.Plotter.convert_to_mp4(filename)
