### 2026-07-11 91f1308
Update TimeSlices.py
