+++ examples/02-Simulating_Hughes_1/data/Mp4converter.py
@@ -3,4 +3,4 @@ from pathlib import Path
 from hughes2d import Plotter
 
 filename = str(Path(__file__).parent / "RuGrandmont")
-Plotter.convertToMP4(filename, limits=[[1,13],[-11,-3]])
+Plotter.convert_to_mp4(filename, limits=[[1,13],[-11,-3]])
