### 2026-07-11 11ddc65
Update DXFImport.py
