### 2026-07-11 c1ca4a3
Update definingZones.py
