+++ examples/06-File_handling/dxf/definingZones.py
@@ -3,5 +3,5 @@ from hughes2d import NonConvexDomain
 filename = "examples/06-File_handling/dxf/config_simple_zones.dxf"
 
 MyDomain = NonConvexDomain()
-MyDomain.importFromDXF(filename)
+MyDomain.import_from_dxf(filename)
 MyDomain.show()
