### 2026-07-11 d058c24
Bump version to 1.1.6
