+++ pyproject.toml
@@ -1,7 +1,7 @@
 
 [project]
 name = "hughes2d"
-version = "1.1.5"
+version = "1.1.6"
 description = "A python package for simulating macroscopic pedestrian evacuations."
 readme = "README.md"
 requires-python = ">=3.11"
