### 2026-07-11 07468ae
Update auto-release.yml
