### 2026-07-11 86b31ec
Update auto-release.yml
