+++ .github/workflows/auto-release.yml
@@ -35,9 +35,10 @@ jobs:
         id: check_release
         run: |
           CURRENT_COMMIT=$(git rev-parse HEAD)
+          CURRENT_COMMIT_TAG=$(git tag --points-at $CURRENT_COMMIT)
           echo "Current release: ${{ steps.get_current_release.outputs.tag_name }}"
-          echo "Current commit: $CURRENT_COMMIT"
-          if [ ${{ steps.get_current_release.outputs.tag_name }} = $CURRENT_COMMIT ] then
+          echo "Current commit: $CURRENT_COMMIT_TAG"
+          if [ ${{ steps.get_current_release.outputs.tag_name }} = $CURRENT_COMMIT_TAG ] then
             echo "already_released=true" >> $GITHUB_OUTPUT
             echo "This commit already has a released tag. Will skip release."
             git tag --points-at $CURRENT_COMMIT
