### 2026-07-11 b904c94
Update auto-release.yml
