### 2026-07-11 b98c8a4
Update auto-release.yml
