### 2026-07-11 9b03ac0
Update auto-release.yml
