### 2026-07-11 8c71296
Update auto-release.yml
