### 2026-07-11 7220991
Update auto-release.yml
