### 2026-07-11 d96fc3c
Update auto-release.yml
