+++ .github/workflows/auto-release.yml
@@ -36,11 +36,11 @@ jobs:
         run: |
           if git tag --points-at HEAD | grep -q "^v"; then
             echo "git_head_valid=true" >> $GITHUB_OUTPUT
-            echo "git is not pointing to a version."
-          else
-            echo "git_head_valid=false" >> $GITHUB_OUTPUT
             echo "git is pointing to the version:"
             git tag --points-at HEAD
+          else
+            echo "git_head_valid=false" >> $GITHUB_OUTPUT
+            echo "git is not pointing to a version. Will skip release."
           fi
       
       - name: Check if already released
