### 2026-07-11 f78eb57
Bump version to 1.1.7
