### 2026-07-20 268d5da
Bump version to 1.1.8
