### 2026-07-20 c8607d3
partial edits in paper.md
