### 2026-07-20 7d2dfdf
suggested edits in paper.md
