+++ paper.md
@@ -25,31 +25,28 @@ bibliography: paper.bib
 
 # State of the field
 
-The mathematical modeling of pedestrian crowd is a rapidly developing topic since a few decades. There exist multiple software packages for simulating crowds of pedestrians -- both open source [@vadere;@JuPedSim;@UMANS;@Cromosim;@CrowdWalk] and proprietary (PTV viswalk, MassMotion). We defer to the [awesome-crowdynamics repository](https://github.com/pozapas/awesome-crowdynamics) for an exhaustive list of the available packages. However, to the best of our knowledge, all these developments deal with microscopic simulations, meaning that a finite -potentially huge- number of particles are simulated. We propose here a Python package for macroscopic simulations of pedestrian evacuations, specifically for the seminal Hughes pedestrian flow model, where the crowd is modeled as a fluid density and not as a finite number of particles.
+The mathematical modeling of a pedestrian crowd is a rapidly developing topic over the past few decades. Multiple software packages exist for simulating crowds of pedestrians -- both open source [@vadere;@JuPedSim;@UMANS;@Cromosim;@CrowdWalk] and proprietary (PTV viswalk, MassMotion). We defer to [awesome-crowdynamics repository](https://github.com/pozapas/awesome-crowdynamics) for an exhaustive list of the available packages. However, to the best of our knowledge, all these developments deal with microscopic simulations, meaning that a finite, potentially huge, number of particles are simulated. Here, we have developed a Python package for macroscopic simulations of pedestrian evacuations, specifically for the seminal Hughes pedestrian flow model, where the crowd is modeled as a fluid density and not as a finite number of particles.
 
 
 # Statement of need
 
-The Hughes model has been thoroughly studied during the last two decades [see @survey], but there exists, at the moment, no general mathematical result of existence of solutions in 2D for this model. While example numerical solutions to the Hughes model can be found in literature [e.g., @Goatin2014], these are not supported by a publicly available and reusable codebase. The present package is aimed at providing a reliable and open-source solution to approximate the behavior of the Hughes model.
+The Hughes model has been thoroughly studied during the last two decades [see @survey], but at the moment, no general mathematical result of existence of solutions in 2D for this model exists. While example numerical solutions to the Hughes model can be found in literature [e.g., @Goatin2014], these are not supported by a publicly available and reusable codebase. This package is aimed at providing a reliable and open-source solution to approximate the behavior of the Hughes model.
 We also hope that this package will help formulating conjectures in the future.
 
 
 # Software design
 
-The `hughes2d` package provides a way to produce simulation of macroscopic pedestrian models without having to learn the mathematical theory related to the underlying PDEs. Indeed, for the Hughes model, particular attention must be dedicated to the choice of numerical schemes: an adapted finite volume scheme for discontinuous scalar conservation law and a fast marching algorithm for the eikonal equation. In fact, the design choices and numerical schemes implemented here result from the collaborations and discussions with many mathematicians (B. Andreianov, V. Perrollaz, F. Peru and A. Sylla to name a few). Additionally, instead of the common choice of square meshes, we extend the range of applications by employing a numerical scheme working on a triangular grid.
[…]
