### 2026-07-20 ec28b7d
minor change in paper.bib
