+++ paper.bib
@@ -101,7 +101,7 @@ doi = {10.17815/cd.2019.21}
 }
 
 @software{JuPedSim,
-author = {Chraibi, Mohcine and Kratz, Kai and Schrödter, Tobias and {The JuPedSim Development Team}},
+author = {Chraibi, Mohcine and Kratz, Kai and Schrödter, Tobias and {the JuPedSim Development Team}},
 title = {{JuPedSim}},
 repository-code = {https://github.com/PedestrianDynamics/jupedsim},
 version = {v1.1.0},
