### 2026-07-20 bac009b
Create CITATION.cff
